+++ src/onegov.ballot/onegov/ballot/collection.py
@@ -1,14 +1,18 @@
 from onegov.ballot.models import Ballot, Vote
-from sqlalchemy import desc
+from sqlalchemy import cast, desc, extract, Integer
 
 
 class VoteCollection(object):
 
-    def __init__(self, session):
+    def __init__(self, session, year=None):
         self.session = session
+        self.year = year
 
     def query(self):
         return self.session.query(Vote)
+
+    def for_year(self, year):
+        return self.__class__(self.session, year)
 
     def get_latest(self):
         """ Returns the votes with the latest date. """
@@ -22,12 +26,32 @@ class VoteCollection(object):
         else:
             return self.by_date(latest_date)
 
+    def get_years(self):
+        """ Returns a list of years for which there are votes. """
+
+        query = self.query()
+        query = query.with_entities(cast(extract('year', Vote.date), Integer))
+        query = query.order_by(desc(Vote.date))
+
+        return list(reversed(sorted((set(r[0] for r in query.all())))))
+
     def by_date(self, date):
         """ Returns the votes on the given date. """
 
         query = self.query()
         query = query.filter(Vote.date == date)
-        query = query.order_by(Vote.domain, Vote.id)
+        query = query.order_by(Vote.domain, Vote.title)
+
+        return query.all()
+
+    def by_year(self, year=None):
+        """ Returns the votes for the current/given year. """
+
+        year = year or self.year
+
+        query = self.query()
+        query = query.filter(extract('year', Vote.date) == year)
+        query = query.order_by(Vote.date, Vote.domain, Vote.title)
 
         return query.all()