+++ src/onegov.ballot/onegov/ballot/tests/test_collection.py
@@ -54,3 +54,49 @@ def test_get_latest(session):
 
     # sort by domain, then by date
     assert [v.title for v in collection.get_latest()] == ['latest']
+
+
+def test_get_years(session):
+    session.add(Vote(
+        title="latest",
+        domain='federation',
+        date=date(2015, 6, 14)
+    ))
+    session.add(Vote(
+        title="older",
+        domain='federation',
+        date=date(2015, 3, 14)
+    ))
+    session.add(Vote(
+        title="even-older",
+        domain='canton',
+        date=date(2013, 6, 12)
+    ))
+
+    session.flush()
+
+    assert VoteCollection(session).get_years() == [2015, 2013]
+
+
+def test_by_years(session):
+    session.add(Vote(
+        title="latest",
+        domain='federation',
+        date=date(2015, 6, 14)
+    ))
+    session.add(Vote(
+        title="older",
+        domain='canton',
+        date=date(2014, 6, 12)
+    ))
+
+    session.flush()
+
+    votes = VoteCollection(session, year=2015)
+    assert len(votes.by_year()) == 1
+    assert votes.by_year()[0].title == "latest"
+
+    assert len(votes.by_year(2014)) == 1
+    assert votes.by_year(2014)[0].title == "older"
+
+    assert len(votes.by_year(2013)) == 0